requirements.txt
# Core dependencies
scrapling[fetchers]>=0.2.9
pydantic>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM Integration
ollama>=0.1.0
openai>=1.0.0
anthropic>=0.3.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-asyncio>=0.23.0

# IMPORTANT: After installing, you MUST run:
#   scrapling install
# This downloads browser dependencies required for StealthyFetcher and the Spider framework.
